+++ 03-image.cpp
@@ -795,7 +795,7 @@ private:
         createBuffer(physicalDevice, logicalDevice, buffersize, VK_BUFFER_USAGE_TRANSFER_SRC_BIT, VK_MEMORY_PROPERTY_HOST_VISIBLE_BIT | VK_MEMORY_PROPERTY_HOST_COHERENT_BIT, stagingBuffer, stagingBufferMemory);
         void *data;
         vkMapMemory(logicalDevice, stagingBufferMemory, 0, buffersize, 0, &data);
-        memcpy(data, vertices.data(), buffersize);
+        memcpy(data, vertices.data(), static_cast<size_t>(buffersize));
         vkUnmapMemory(logicalDevice, stagingBufferMemory);
 
         //use device local buffer is fastest
@@ -818,7 +818,7 @@ private:
         createBuffer(physicalDevice, logicalDevice, buffersize, VK_BUFFER_USAGE_TRANSFER_SRC_BIT, VK_MEMORY_PROPERTY_HOST_VISIBLE_BIT | VK_MEMORY_PROPERTY_HOST_COHERENT_BIT, stagingBuffer, stagingBufferMemory);
         void *data;
         vkMapMemory(logicalDevice, stagingBufferMemory, 0, buffersize, 0, &data);
-        memcpy(data, indices.data(), buffersize);
+        memcpy(data, indices.data(), static_cast<size_t>(buffersize));
         vkUnmapMemory(logicalDevice, stagingBufferMemory);
 
         //use device local buffer is fastest
@@ -850,7 +850,7 @@ private:
             createBuffer(physicalDevice, logicalDevice, buffersize, VK_BUFFER_USAGE_UNIFORM_BUFFER_BIT, VK_MEMORY_PROPERTY_HOST_COHERENT_BIT | VK_MEMORY_PROPERTY_HOST_VISIBLE_BIT, uboMVPBuffers[i], uboMVPBufferMemorys[i]);
             void *data;
             vkMapMemory(logicalDevice, uboMVPBufferMemorys[i], 0, buffersize, 0, &data);
-            memcpy(data, &uboMVP, buffersize);
+            memcpy(data, &uboMVP, static_cast<size_t>(buffersize));
             vkUnmapMemory(logicalDevice, uboMVPBufferMemorys[i]);
         }
     }
@@ -882,7 +882,7 @@ private:
 
         void *data;
         vkMapMemory(logicalDevice, stagingBufferMemory, 0, imageSize, 0, &data);
-        memcpy(data, pixels, imageSize);
+        memcpy(data, pixels, static_cast<size_t>(imageSize));
         vkUnmapMemory(logicalDevice, stagingBufferMemory);
 
         stbi_image_free(pixels);
@@ -965,7 +965,7 @@ private:
     }
     void createImageTextureView()
     {
-        testImageView = createImageView(logicalDevice, testImage, VK_IMAGE_VIEW_TYPE_2D, VK_FORMAT_R8G8B8A8_UNORM,VK_IMAGE_ASPECT_COLOR_BIT);
+        testImageView = createImageView(logicalDevice, testImage, VK_IMAGE_VIEW_TYPE_2D, VK_FORMAT_R8G8B8A8_UNORM, VK_IMAGE_ASPECT_COLOR_BIT);
         sampler = createSampler(logicalDevice, VK_FILTER_LINEAR, VK_FILTER_LINEAR, VK_SAMPLER_MIPMAP_MODE_LINEAR);
     }
 };
